requirements.txt
autopep8
certifi
chardet
defusedxml
Django
django-allauth
django-countries
django-crispy-forms
django-debug-toolbar
idna
oauthlib
pep8
Pillow
pycodestyle
python-decouple
python3-openid
pytz
requests
requests-oauthlib
sqlparse
# stripe==2.27.0
urllib3
django-paypal
paypalrestsdk
djangorestframework
crispy-bootstrap4
